Switch to long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
 
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[] {6, 4, 3});
-        int max = salesManager.max();
+        SalesManager salesManager = new SalesManager(new long[] {6, 4, 3});
+        long max = salesManager.max();
         System.out.println(max);
     }
 }
